Showing variables at solution view

+++ src/app/simplex.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { Result, Tabloid } from 'src/native/simplex';
 import { MatricialForm } from './parser/models/matricial-form';
 
@@ -10,18 +10,22 @@ export class SimplexService {
 
   constructor() { }
 
-  evaluate(mat: MatricialForm) {
-    return new Observable<{ answear: Result, steps: Tabloid[] }>(observer => {
-      const worker = new Worker('./simplex.worker', { type: 'module' });
-      worker.onmessage = ({ data }) => {
-        observer.next(data);
-        observer.complete();
-        worker.terminate();
-      };
-      worker.postMessage(mat);
-      return () => {
-        worker.terminate();
-      };
-    });
+  evaluate(mat: MatricialForm | null) {
+    if (mat) {
+      return new Observable<{ answear: Result, steps: Tabloid[] }>(observer => {
+        const worker = new Worker('./simplex.worker', { type: 'module' });
+        worker.onmessage = ({ data }) => {
+          observer.next(data);
+          observer.complete();
+          worker.terminate();
+        };
+        worker.postMessage(mat);
+        return () => {
+          worker.terminate();
+        };
+      });
+    } else {
+      return of(null);
+    }
   }
 }
